+++ src/util.hpp
@@ -18,7 +18,7 @@ bits256 hexadecimal_digest_to_bits256(std::string digest_hex_str);
 bits64 hexadecimal_value_to_bits64(std::string value_hex_str);
 
 std::vector<bool> convert_int_to_binary(int x);
-std::vector<bool> address_bits_from_address(int address, int tree_depth);
+std::vector<bool> address_bits_from_address(int address, size_t tree_depth);
 
 } // libzeth
